multilib packages depend on their regular counterparts
Multilib packages are usually stripped from everything that could conflict with
their regular counterparts (bin, include, share) and thus depend on them.
Currently namcap warns that they are not needed. There are of course a few
exceptions, so instead of adding the regular package in smartdepend earlier in
the script, skipping the warning seems to be the way to go.

Signed-off-by: Kyle Keen <keenerd@gmail.com>