parsepkgbuild.sh: fix detection of split pkgbuilds with empty pkgbase
For split pkgbuilds the pkgbase variable is optional. If not specified,
the first element in the pkgname array is used. Currently parsepkgbuild.sh
fails to detect split pkgbuilds if pkgbase is not specified. Therefore the
split pkgbuild rules are not applied to such pkgbuilds.

Instead of testing if pkgbase is set, test if the pkgname array has more
than one element. If so, set pkgbase accordingly.

Signed-off-by: Michael Straube <michael.straube at posteo.de>
Signed-off-by: Kyle Keen <keenerd@gmail.com>