recognize arch-specific variables even if packages is for only one architecture
Architecture-specific variables should be useable as soon as a package
is not built for 'any' architecture. With the deprecation of i686, this
becomes relevant, as many packages are now only built for x86_64.
However, the source (and checksums) may be _not_ architecture-agnostic,
e.g. should _not_ be listed as source= but source_x86_64=.

This patch silences the extra-var-begins-without-underscore message
which is currently emitted in that case.

Signed-off-by: Erich Eckner <arch at eckner.net>
Signed-off-by: Kyle Keen <keenerd@gmail.com>