Add rule to report on unneeded dependencies from hooks
Just as it is unnecessary to run certain commands in a post_upgrade
script, if a hook already runs it, so too is it usually unnecessary to
depend on the package that installs the hook.

Sometimes these dependencies are left in even when the install script is
removed, because people lack understanding in the true subtleties of
hooks and the commands they run -- but the commands do not depend on the
list of files which trigger them in order to function, and are fully
operative when installed later. In fact, they will already be required
by the software that uses the results.

e.g. desktop-file-utils and shared-mime-info provide hooks to compile
the mime and desktop file databases, but upon first installing the hook,
a full database will already be compiled. The hook does not need to be
triggered by the filepath, and thus does not need to be a dependency of
every package that provides a desktop or mime file. Instead, desktop
environments which need a mime/desktop file database will depend on the
package and thereby ensure the database exists when needed.

Signed-off-by: Eli Schwartz <eschwartz at archlinux.org>
Signed-off-by: Kyle Keen <keenerd@gmail.com>