remove pointless mkdir -p in parsepkgbuild
As far as I can tell, this was never used.

In commit 9b56647abc8dc33bf138d4b260b103c8d9265b7d the parsepkgbuild
script was switched from defining a PATH as a directory in /tmp, to
actually passing in a PATH=/dummy to `env -i`... which makes sense as
the defined PATH did not used to be used. However, even before that,
it's not safe to use mkdir -p in order to define a restricted shell
without access to the PATH, as it's not guaranteed to be owned by the
creator.

In any event, the time for mkdir -p is long past.

Signed-off-by: Eli Schwartz <eschwartz at archlinux.org>
Signed-off-by: Kyle Keen <keenerd@gmail.com>